// src/types/coordinators.ts
export interface CoordinatorMember {
  id: string;
  name: string;
  role: string;
  department: string;
  year?: string;
  image: string;
  designation?: string;
}

export interface FacultyCoordinatorMember {
  id: string;
  name: string;
  role: string;
  designation: string;
  department: string;
  image: string;
}